% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tba_ranking.R
\name{calculate_tba_ranking}
\alias{calculate_tba_ranking}
\title{Calculate the TBA rankings from qualifying match scores for the current season}
\usage{
calculate_tba_ranking(data)
}
\arguments{
\item{data}{tibble of match information - it is assumed that each line corresponds
to the score breakdown of one team in one event's match.}
}
\description{
TBA rankings can be calculated from partial information, for example before all of the qualifying
matches are played or scores are posted.
}
\examples{
# example code
\dontrun{
# get all matches for one event:
clash_matches <- get_records("event/2024cttd/matches")
# get the score breakdown for each of the teams in each match
clash_details <- get_match_details(clash_matches)
# now get the rankings:
calculate_tba_ranking(clash_details)
}
}
