% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tba_ranking.R
\name{calculate_tba_ranking_2024}
\alias{calculate_tba_ranking_2024}
\title{(Re-calculate) the TBA rankings from qualifying match scores}
\usage{
calculate_tba_ranking_2024(data)
}
\arguments{
\item{data}{tibble of match information - it is assumed that each line corresponds
to the score breakdown of one team in one event's match.}
}
\description{
Note that rankings for historic events can be sourced from TBA with a call of the form
\code{get_records("/event/{event_key}/rankings")$rankings}
}
\examples{
# example code
# get all matches for one event:
clash_matches <- get_records("event/2024cttd/matches")
# get the score breakdown for each of the teams in each match
clash_details <- get_match_details(clash_matches)
# now get the rankings:
calculate_tba_ranking_2024(clash_details)
}
