% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{check_key}
\alias{check_key}
\title{Check the API key}
\usage{
check_key(key)
}
\arguments{
\item{key}{character string. Authorization key for read access to TBA's APIv3 (2025).}
}
\value{
Boolean. TRUE, for successful status check.
}
\description{
Run a status check on the given key. Returns \code{TRUE} when the key is valid.
}
