% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_teams.R
\name{compare_teams}
\alias{compare_teams}
\title{Get a parallel coordinate plot for an event}
\usage{
compare_teams(event_code, ..., color = "Pick 1", verbose = TRUE)
}
\arguments{
\item{event_code}{character string of an event}

\item{...}{set of variables that should be used in the comparison of teams}

\item{color}{character value, one of "Captain", "Pick 1", "Pick 2", "Pick 3". Colors all
identified members of an Alliance up to the specified level. Use \code{NA} for no coloring.}

\item{verbose}{logical value. Should the function provide messages during runtime? Defaults to TRUE.}
}
\value{
a parallel coordinate plot
}
\description{
Show the parallel coordinate plot of an event based on the qualification matches
and the TBA ranking.
}
\details{
Note that this function is a wrapper of  \code{tba} functionality and is therefore
quite opinionated. In case you want to make changes,
you would need to work with the basic \code{tba} functions.
Assumptions/Decisions:
\itemize{
\item comparisons are based on qualifying matches only.
}
}
\examples{
# just get the plot
compare_teams("2024iacf", score)
}
