% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{get_api_key}
\alias{get_api_key}
\alias{set_api_key}
\title{Read and set the API key for read access to TBA}
\usage{
get_api_key()

set_api_key(key, check = TRUE, overwrite_check = TRUE)
}
\arguments{
\item{key}{character string. Authorization key for read access to TBA's APIv3 (2025).}

\item{check}{Boolean. Should the key be checked for validity?}

\item{overwrite_check}{Boolean. Should user be asked before an existing key is overwritten?}
}
\value{
\code{get_api_key} returns the key as an invisible object. \code{set_api_key} sets the specified key as environment variable.
}
\description{
\code{get_api_key} checks whether the key for the API to The Blue Aliance (TBA) is available
in the system environment. The function will warn the user in case the key is not found and
will provide instructions on how to set the key.
In order to access the API, the user will have to sign up for a (free) account with \url{https://www.thebluealliance.com/account}.
The functions in this package expect access to TBA's read APIv3. Once created,
call the function \code{set_api_key} with the string found under \code{X-TBA-Auth-Key}.
\code{set_api_key} makes the specified key available as a variable in the system environment for the current session.
}
