% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_details.R
\name{get_match_details}
\alias{get_match_details}
\title{Reformat match information into an approximate wide form}
\usage{
get_match_details(data)
}
\arguments{
\item{data}{tibble of match records,  assumed to be in the form of a result
from a call to \verb{get_records(/event/\{event_key\}/matches)}.}
}
\value{
tibble of scores and score breakdowns for each match and each team
}
\description{
The records from TBA come in the form of one line for each match, where the
two alliances and their corresponding score breakdowns are in list variables.
The \code{get_match_details} function returns the score breakdown for each team and each match in the
competition, i.e. the new dataset will have about 6 times the number of rows of the original.
HH: still need to test: what happens, when there is a surrogate team used in a match?
}
\examples{
clash_matches <- get_records("event/2024iawes/matches")
clash_details <- get_match_details(clash_matches)
clash_matches <- get_records("event/2025iacf/matches")
clash_details <- get_match_details(clash_matches)
}
