% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ranking_by_diff.R
\name{get_ranking_by_diff}
\alias{get_ranking_by_diff}
\title{Rank teams in the qualifying matches of an event by a scoring variable}
\usage{
get_ranking_by_diff(data, ..., method = "normal")
}
\arguments{
\item{data}{tibble of match information - it is assumed that each line corresponds
to the score breakdown of one team in one event's match.}

\item{...}{numeric variables measuring aspects a team's performance (in qualifying matches).}

\item{method}{character which model should be calculated? Either \code{normal} or \code{loglinear}.}
}
\value{
tibble of the ranked teams and the number of matches (\code{n}) that the ranking is based on.
}
\description{
All teams are ranked (if possible) by their contribution to the difference in score in (qualifying) matches of the First Robotics Compettition.
These rankings are generally based on matches from the qualifying round and provide teams with
more information to pick other teams when they form alliances.
}
\details{
In order to model each team's contribution to the score difference, a wide data set has to be constructed, with each
team as a variable, and each row as one of the alliances in one match with the difference in score as the dependent variable.
An alliance in a match consists of exactly three robots. This requirement is checked before modelling
and if not fulfilled, results in a warning.
A normal model is then fitted, or, if \code{method = "loglinear"} a log-linear model is fitted, if all values in the scoring variable are non-negative.
The coefficients of this model can be interpreted as the average (log) contribution of that
team to the difference in score of a match. Based on this information, an expected score difference can be calculated for an alliance.
}
\examples{
# example code
# get all matches for one event:
matches <- get_records("event/2024cttd/matches")
# get the score breakdown for each of the teams in each match
match_details <- get_match_details(matches)

library(dplyr, quietly=TRUE)
# now get the contribution to the score:
match_details \%>\% filter(comp_level == "qm") \%>\% # get the qualifying matches
  get_ranking_by_diff(score)

# additionally get the team contributions to the score without counting the opponents' fouls:
match_details \%>\% filter(comp_level == "qm") \%>\% 
  get_ranking_by_diff(score, score-foulPoints)
}
