% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Get a dataframe with records from TBA}
\usage{
get_records(call)
}
\arguments{
\item{call}{call string from \code{list}  https://www.thebluealliance.com/apidocs/v3#data-tag-list}
}
\value{
tibble (or list) of the content.
}
\description{
The Blue Alliance API v3 is specified at https://www.thebluealliance.com/apidocs/v3.
The function \code{get_records} takes a \code{call} to any of the TBA lists under \code{list} and returns, if possible, all records in a data frame.
In case the results can not be converted to a data frame, they are returned as list.
}
\examples{
events <- get_records("events/2024")
events[grep("Corn", events$name),] # yep, there is an event with 'Corn' in it

table(events$state_prov)

# now draw a map of all events in North America
require(ggplot2)
require(dplyr)
events \%>\% slice(grep("America", timezone)) \%>\% 
  filter(lat > 20) \%>\% # sorry, South America
  ggplot(aes(x = lng, y = lat)) + geom_point() + 
  coord_equal()

}
