% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_ranking.R
\name{tba_ranking}
\alias{tba_ranking}
\title{Get team rankings for an event}
\usage{
tba_ranking(event_code)
}
\arguments{
\item{event_code}{character string for an event -  usually an eight-character string consisting of a four-digit year followed by an acronym for the event, such as 'cttd' for "Cow Town Throw Down".}
}
\value{
tibble of the rankings for all teams that participated in the event
}
\description{
Get team rankings for an event
}
\examples{
tba_ranking("2024cttd") # same team ranking as in example for `calc_ranking_tba_2024`
tba_ranking("2025iacf") 
}
