% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_ranking.R
\name{tba_ranking}
\alias{tba_ranking}
\title{Get team rankings for an event}
\usage{
tba_ranking(event_code, season)
}
\arguments{
\item{event_code}{character string for an event}

\item{season}{four-digit year}
}
\value{
tibble of the rankings for all teams that participated in the event
}
\description{
Get team rankings for an event
}
\examples{
tba_ranking("cttd", 2024) # same team ranking as in example for `calc_ranking_tba_2024`
}
